pyproject.toml
[project]
name = "pysec-clients"
version = "0.1.0"
description = "Monorepo for security vendor API clients, e.g. Defender, CrowdStrike, Tenable, and more to come."
readme = "README.md"
license = "AGPL-3.0"
authors = [
    { name = "Proxay", email = "28168887+ProxayFox@users.noreply.github.com" },
]
requires-python = ">=3.14"
dependencies = ["mde-client[arrow,polars]"]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "checkov>=3.1.47",
    "deptry>=0.25.1",
    "pre-commit>=4.5.1",
    "psutil>=7.2.2",
    "pyarrow-stubs>=20.0.0.20251215",
    "pyinstrument>=5.1.2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.11",
    "scalene>=2.2.1",
    "scalene-mcp>=0.1.0",
    "ty>=0.0.32",
    "types-defusedxml>=0.7.0.20260408",
    "types-pyyaml>=6.0.12.20260408",
    "vulture>=2.16",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.6",
    "pymdown-extensions>=10.21.2",
]

[tool.uv.workspace]
members = ["src/*"]

[tool.uv.sources]
mde-client = { workspace = true }

[tool.ty.environment]
python = "./.venv"
root = ["./src", "./test"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "test"]
extraPaths = ["src"]
